Fix int8 matrix multiplication
1. Correct feature grain size which allows Matrix A size to be 544x544.
2. Correct cpu calculation for int8.

+++ include/npu_cna.h
@@ -29,7 +29,7 @@ typedef struct npu_cna_desc {
   uint8_t in_precision;       // 0x100C
   uint8_t proc_precision;     // 0x100C
   uint8_t kernel_groups;      // 0x1010
-  uint8_t feature_grains;     // 0x1010
+  uint16_t feature_grains;    // 0x1010
   uint8_t conv_y_stride;      // 0x1014
   uint8_t conv_x_stride;      // 0x1014
   uint16_t datain_width;      // 0x1020
